requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
PyPDF2
chromadb
faiss-cpu==1.7.4
langchain_google_genai
youtube_transcript_api
pathlib
flask